Move analysis persistence into package and store checkpoints in refs
