+++ diagram_analysis/__init__.py
@@ -1,4 +1,4 @@
 from .diagram_generator import DiagramGenerator
-from .manifest import AnalysisManifest, load_manifest, save_manifest
+from .persistence.manifest import AnalysisManifest, load_manifest, save_manifest
 from .run_context import RunContext
 from agents.llm_config import configure_models
