+++ diagram_analysis/persistence/version.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class Version(BaseModel):
+    commit_hash: str
+    code_boarding_version: str
